+++ sonaranalyzer-dotnet/src/SonarAnalyzer.Common/Rules/TooManyParametersBase.cs
@@ -18,11 +18,10 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System.Collections.Generic;
+using System.Linq;
 using Microsoft.CodeAnalysis;
 using SonarAnalyzer.Common;
 using SonarAnalyzer.Helpers;
-using System.Collections.Immutable;
 
 namespace SonarAnalyzer.Rules
 {
@@ -31,17 +30,19 @@ namespace SonarAnalyzer.Rules
         where TParameterListSyntax : SyntaxNode
     {
         protected const string DiagnosticId = "S107";
-        protected const string MessageFormat = "{2} has {1} parameters, which is greater than the {0} authorized.";
+        protected const string MessageFormat = "{0} has {1} parameters, which is greater than the {2} authorized.";
 
-        protected abstract DiagnosticDescriptor Rule { get; }
         protected abstract TSyntaxKind[] SyntaxKinds { get; }
         protected abstract GeneratedCodeRecognizer GeneratedCodeRecognizer { get; }
-        protected abstract Dictionary<TSyntaxKind, string> Mapping { get; }
-        protected abstract TSyntaxKind ParentType(TParameterListSyntax parameterList);
+
+        protected abstract string UserFriendlyNameForNode(SyntaxNode node);
+
         protected abstract int CountParameters(TParameterListSyntax parameterList);
+
         protected abstract bool CanBeChanged(SyntaxNode node, SemanticModel semanticModel);
 
         private const int DefaultValueMaximum = 7;
+
         [RuleParameter("max", PropertyType.Integer, "Maximum authorized number of parameters", DefaultValueMaximum)]
         public int Maximum { get; set; } = DefaultValueMaximum;
 
@@ -56,14 +57,42 @@ namespace SonarAnalyzer.Rules
 
                     if (parametersCount > Maximum &&
                         parameterListNode.Parent != null &&
-                        CanBeChanged(parameterListNode.Parent, c.SemanticModel) &&
-                        Mapping.TryGetValue(ParentType(parameterListNode), out var declarationName))
+                        CanBeChanged(parameterListNode.Parent, c.SemanticModel))
                     {
-                        c.ReportDiagnosticWhenActive(Diagnostic.Create(Rule, parameterListNode.GetLocation(),
-                            Maximum, parametersCount, declarationName));
+                        c.ReportDiagnosticWhenActive(Diagnostic.Create(SupportedDiagnostics[0], parameterListNode.GetLocation(),
+                            UserFriendlyNameForNode(parameterListNode.Parent), parametersCount, Maximum));
                     }
                 },
                 SyntaxKinds);
+        }
+
+        protected bool VerifyCanBeChangedBySymbol(SyntaxNode node, SemanticModel semanticModel)
+        {
+            var declaredSymbol = semanticModel.GetDeclaredSymbol(node);
+            var symbol = semanticModel.GetSymbolInfo(node).Symbol;
+            if (declaredSymbol == null && symbol == null)
+            {
+                // no information
+                return false;
+            }
+
+            if (symbol != null)
+            {
+                // Not a declaration, such as Action
+                return true;
+            }
+
+            if (declaredSymbol.IsExtern &&
+                declaredSymbol.IsStatic &&
+                declaredSymbol.GetAttributes(KnownType.System_Runtime_InteropServices_DllImportAttribute).Any())
+            {
+                // P/Invoke method is defined externally.
+                // Do not raise
+                return false;
+            }
+
+            return declaredSymbol.GetOverriddenMember() == null &&
+                   declaredSymbol.GetInterfaceMember() == null;
         }
     }
 }
